add password validation
