+++ components/index.ts
@@ -16,6 +16,7 @@ export { default as UserReviewsView } from "./views/UserReviewsView";
 export { default as ViewUserProfile } from "./views/ViewUserProfile";
 export { default as ViewProSignUp } from "./views/ViewProSignUp";
 export { default as ViewRequestPasswordReset } from "./views/ViewRequestPasswordReset";
+export { default as ViewResetPassword } from "./views/ViewResetPassword";
 
 // Cards
 export { default as AddReviewModal } from "./cards/AddReviewModal";
